+++ custom_components/lsc_tuya_doorbell/const.py
@@ -227,6 +227,8 @@ PLATFORMS = [
     "select",
     "number",
     "camera",
+    "image",
+    "button",
 ]
 
 # SD Card status mapping